made a separate file for get_search_term_query
