+++ firstvoices/backend/search/utils/query_builder_utils.py
@@ -14,14 +14,6 @@ from backend.search.indices.dictionary_entry_document import (
 )
 from backend.search.utils.constants import VALID_DOCUMENT_TYPES
 
-BASE_BOOST = 1.0  # default value of boost
-EXACT_MATCH_PRIMARY_BOOST = 5
-EXACT_MATCH_SECONDARY_BOOST = 4
-FUZZY_MATCH_PRIMARY_BOOST = 3
-FUZZY_MATCH_SECONDARY_BOOST = 2
-EXACT_MATCH_OTHER_BOOST = 1.5
-FUZZY_MATCH_OTHER_BOOST = BASE_BOOST
-
 
 class SearchDomains(Enum):
     BOTH = "both"
@@ -64,180 +56,6 @@ def get_types_query(types):
         return Q(~Q("match", type=TypeOfDictionaryEntry.WORD))
     else:
         return None
-
-
-def get_search_term_query(search_term, domain):
-    # Exact matching has a higher boost value, then fuzzy matching for both title and translation fields
-
-    # Primary fields
-    exact_match_primary_language_query = Q(
-        {
-            "match_phrase": {
-                "primary_language_search_fields": {
-                    "query": search_term,
-                    "slop": 3,  # How far apart the terms can be in order to match
-                    "boost": EXACT_MATCH_PRIMARY_BOOST,
-                }
-            }
-        }
-    )
-    exact_match_primary_translation_query = Q(
-        {
-            "match_phrase": {
-                "primary_translation_search_fields": {
-                    "query": search_term,
-                    "slop": 3,
-                    "boost": EXACT_MATCH_PRIMARY_BOOST,
-                }
-            }
-        }
-    )
-    fuzzy_match_primary_language_query = Q(
-        {
-            "fuzzy": {
-                "primary_language_search_fields": {
-                    "value": search_term,
-                    "fuzziness": "2",  # Documentation recommends "AUTO" for this param
-                    "boost": FUZZY_MATCH_PRIMARY_BOOST,
-                }
-            }
-        }
-    )
-    fuzzy_match_primary_translation_query = Q(
-        {
-            "fuzzy": {
-                "primary_translation_search_fields": {
-                    "value": search_term,
-                    "fuzziness": "2",
-                    "boost": FUZZY_MATCH_PRIMARY_BOOST,
-                }
-            }
-        }
-    )
-
-    # Secondary fields
-    exact_match_secondary_language_query = Q(
-        {
-            "match_phrase": {
-                "secondary_language_search_fields": {
-                    "query": search_term,
-                    "slop": 3,
-                    "boost": EXACT_MATCH_SECONDARY_BOOST,
-                }
-            }
-        }
-    )
-    exact_match_secondary_translation_query = Q(
-        {
-            "match_phrase": {
-                "secondary_translation_search_fields": {
-                    "query": search_term,
-                    "slop": 3,
-                    "boost": EXACT_MATCH_SECONDARY_BOOST,
-                }
-            }
-        }
-    )
-    fuzzy_match_secondary_language_query = Q(
-        {
-            "fuzzy": {
-                "secondary_language_search_fields": {
-                    "value": search_term,
-                    "fuzziness": "2",
-                    "boost": FUZZY_MATCH_SECONDARY_BOOST,
-                }
-            }
-        }
-    )
-    fuzzy_match_secondary_translation_query = Q(
-        {
-            "fuzzy": {
-                "secondary_translation_search_fields": {
-                    "value": search_term,
-                    "fuzziness": "2",
-                    "boost": FUZZY_MATCH_SECONDARY_BOOST,
-                }
-            }
-        }
-    )
-
-    # Other fields
-    exact_match_other_language_query = Q(
-        {
-            "match_phrase": {
-                "other_language_search_fields": {
-                    "query": search_term,
-                    "slop": 3,
-                    "boost": EXACT_MATCH_OTHER_BOOST,
-                }
-            }
-        }
-    )
-    exact_match_other_translation_query = Q(
-        {
-            "match_phrase": {
-                "other_translation_search_fields": {
-                    "query": search_term,
-                    "slop": 3,
-                    "boost": EXACT_MATCH_OTHER_BOOST,
-                }
-            }
-        }
-    )
-    fuzzy_match_other_language_query = Q(
-        {
-            "fuzzy": {
-                "other_language_search_fields": {
-                    "value": search_term,
-                    "fuzziness": "2",
-                    "boost": FUZZY_MATCH_OTHER_BOOST,
-                }
-            }
-        }
-    )
-    fuzzy_match_other_translation_query = Q(
-        {
-            "fuzzy": {
-                "other_translation_search_fields": {
-                    "value": search_term,
-                    "fuzziness": "2",
-                    "boost": FUZZY_MATCH_OTHER_BOOST,
-                }
-            }
-        }
-    )
-
-    subqueries = []
-
-    subquery_domains = {
-        "language": [
-            exact_match_primary_language_query,
-            fuzzy_match_primary_language_query,
-            exact_match_secondary_language_query,
-            fuzzy_match_secondary_language_query,
-            exact_match_other_language_query,
-            fuzzy_match_other_language_query,
-        ],
-        "translation": [
-            exact_match_primary_translation_query,
-            fuzzy_match_primary_translation_query,
-            exact_match_secondary_translation_query,
-            fuzzy_match_secondary_translation_query,
-            exact_match_other_translation_query,
-            fuzzy_match_other_translation_query,
-        ],
-    }
-    subquery_domains["both"] = (
-        subquery_domains["language"] + subquery_domains["translation"]
-    )
-
-    subqueries += subquery_domains.get(domain, [])
-
-    return Q(
-        "bool",
-        should=subqueries,
-        minimum_should_match=1,
-    )
 
 
 def get_site_filter_query(site_id):
